+++ src/TwigComponent/src/DependencyInjection/TwigComponentExtension.php
@@ -150,10 +150,11 @@ final class TwigComponentExtension extends Extension implements ConfigurationInt
         $container->setAlias('console.command.stimulus_component_debug', 'ux.twig_component.command.debug')
             ->setDeprecated('symfony/ux-twig-component', '2.13', '%alias_id%');
 
-        if ($container->getParameter('kernel.debug') && $config['profiler']) {
+        if ($config['profiler']['enabled']) {
             $loader->load('debug.php');
+
             $container->getDefinition('ux.twig_component.data_collector')
-                ->setArgument(2, $config['profiler_dump_components'] ?? true);
+                ->setArgument(2, $config['profiler']['collect_components']);
         }
 
         $loader->load('cache.php');
@@ -217,13 +218,13 @@ final class TwigComponentExtension extends Extension implements ConfigurationInt
                 ->scalarNode('anonymous_template_directory')
                     ->info('Defaults to `components`')
                 ->end()
-                ->booleanNode('profiler')
-                    ->info('Enables the profiler for Twig Component (in debug mode)')
-                    ->defaultValue('%kernel.debug%')
-                ->end()
-                ->booleanNode('profiler_dump_components')
-                    ->info('Dump components in the Twig Component profiler panel')
-                    ->defaultTrue()
+                ->arrayNode('profiler')
+                    ->info('Enables the profiler for Twig Component')
+                    ->canBeEnabled()
+                    ->children()
+                        ->booleanNode('enabled')->defaultValue('%kernel.debug%')->end()
+                        ->booleanNode('collect_components')->info('Collect components instances')->defaultTrue()->end()
+                    ->end()
                 ->end()
                 ->scalarNode('controllers_json')
                     ->setDeprecated('symfony/ux-twig-component', '2.18', 'The "twig_component.controllers_json" config option is deprecated, and will be removed in 3.0.')