+++ src/TwigComponent/CHANGELOG.md
@@ -1,5 +1,10 @@
 # CHANGELOG
 
+## 2.32
+
+-  Add option `profiler.collect_components` to control component data collection
+   in the profiler (enabled in debug mode by default)
+
 ## 2.30
 
 -  Ensure compatibility with PHP 8.5